OVAL: Do not treat empty HyperShift file as an indication that HyperShift is installed
The `installed_app_is_ocp4_{{{ minorversion }}}` check tries to make
sure that hyperShift clusteroperator file is not present, but it does so
by checking the file presence. This does not work, because even for an
empty file, the api-resource-collector saves a file with contents:
```
kube-api-error=NotFound
```
Let's check that the file matches a regex starting with 4 instead.