Creates comments informing the dictionary structure
item.value[1]|int is used in some Ansible tasks when looping through a
dictionary created from /etc/passwd file. It could not be intuitive the
index and values of the resulting dictionary. It is basically formed by
the fields in /etc/passwd, where the first field is the key and the next
6 fields are in the list of values. Each file line is a item in the
dictionary.