OCPBUGS-7307: Remove templating from OCIL in alert_receiver_configured
The alert_receiver_configured rule has instructions/OCIL that tells
the user how to check if an alerting system is setup. The command relies
on a golang templating feature of the kubeclient to fish the appropriate
data out of OpenShift. The Compliance Operator uses the same
templating syntax to resolve variables in OCIL so that it can render
instructions appropriately.

These two template usages conflicted silently when the operator parsed
the datastream, causing the Compliance Operator to omit instructions for
that rule.

This is the only occurrence where the content suggests users leverage
the templating feature, so it should be the only rule that doesn't have
instructions as a result of this problem.

This commit updates the OCIL for that particular rule to use a different
way of getting the data that doesn't rely on the golang text templating
library feature exposed through the kubeclient. It also fixes a trival
spelling mistake in the rule.