Set sshd_strong_kex variable in CIS RHEL7 control
The current CIS Benchmark for RHEL7 requires suggest a different set of
algorithms not present in the current variable options. A new option was
included specific for RHEL7. In the future, we should try to use more
generic variables instead of product specific options but this would
demand more investigation to confirm if it is reasonable in this
context.