Increase robustness of OVAL check
It was included a new test case to cover systems where there is no user
with a password defined. In this case, the remediation is not necessary.
It was also included a filter in the shadow_object to exclude entries
without a valid password. Finally, the file was aligned to the project
Style Guide.