Increase the robustness of OVAL check
The previous check was reporting "unknown" when there is no password
defined in the /etc/shadow file of the checked system. Before, it was
not clear the reason for the unknown result. Therefore, a new test was
included to check if passwords are inexistent in the system. The rule is
no longer reporting "unknown" and the generated reports are more clearly
informing which checks passed or not.