Update ansible in file_permissions_binary_dirs
Add the state argument to the file module. When leaving empty ansible
tries to guess it, and if it detects it is a hard link it will require
an src, which is not given so it'll fail. So setting the state to
'file' it only changes the permissions, as expected

Signed-off-by: Edgar Aguilar <edgar.aguilar@oracle.com>