Update bash remediation
The Bash remediation was using the a macro specific for this rule which
was very complex to be maintained and was not reutilizing existing
macros to avoid code duplication. This macro was removed from the bash
remediation and a much simpler remediation was created using existing
macros shared by other rules. The remediation is aligned to the approach
used in audit_rules_privileged_commands template. Note that some bash
variables are empty but were intentionally kept to preserv the
readability when calling the macro.