Use check-fact-ref in platform element
We will use `<check-fact-ref>` instead of `<fact-ref>` within
the `<platform>` element. The `<fact-ref>` references a `<cpe-item>`
within a CPE dictionary. The `<check-fact-ref>` directly references
the associated OVAL check for the platform.

The effect of this change is that the generated CPE dictionary will not
contain CPE items for CPE AL platforms. Therefore, this change
simiplifies platform definitions in the built content and makes their
processing easier and reduces size of the XML file.

For more details about the check-fact-ref element, please see
sections 5.4 and 6.5 in the CPE 2.3 specification:
https://nvlpubs.nist.gov/nistpubs/Legacy/IR/nistir7698.pdf