Improve Ansible remediation
The Ansible remediation was collecting information in an inefficient
way and repeating similar tasks unnecessarily. The Playbook was
simplified to improve readability and also optimized to be more
efficient. From the ansible_facts is filtered only the mount points
eligible to execute privileged commands. The privileged commands are
search only in these mount points. Finally, the decision to include a
new rule or update an existing rule is left to Ansible lineinfile
module.