Improve version sanity check
In this commit, we improve version sanity check for the platform_package
platform template to make sure that the version string supplied by the
build system is always normalized and contains an epoch. The normal form
with epoch is important to make the version comparison more reliable.
To perform the version check we will now check the variable that is
actually used in the template ("ev_ver") instead of unused variable.