OCPBUGS-8347: Disable the check for scheduler's bind-address option, the option was removed from newer kuber versions
The `scheduler_no_bind_address` option concerns exposing port 10251/TCP
of the scheduler to any other interface than localhost using the
bind-address parameter. The healthz service was running unauthenticated
on that port, so it was important that it is not bound to outside facing
interfaces.

However, the unsecure port option was removed completely in Kube
versions starting with 1.23:
    https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG/CHANGELOG-1.23.md#deprecation
which means that the bind-address parameter is now safe to use.

And because OCP 4.9 is the last one to use kube 1.22:
    https://access.redhat.com/solutions/4870701
and because OCP 4.8 is the oldest supported now, we can restrict this
rule to OCP 4.8 and 4.9.