Remove confused test scenario
The fail test scenario "db_malformed.fail.sh" creates a malformed
AIDE database in /var/lib/aide/aide.db.gz. The name suggests that
the rule should fail because the database is malformed. However,
as we can see from scan results or from oval-graph, the reason
that the rule fails isn't that the database is malformed. Instead,
the reason that the rule fails is that the so-called "new" database
file /var/lib/aide/aide.db.new.gz doesn't exist. In fact, the OVAL
check in this rule doesn't have any way of checking the contents
of the files, it only checks for presence of files.