Fix authselect check task
The a7cdf351071949a488c21de07fa2a341263abea8 caused an Ansible task to
fail by replacing the "ignore_errors" parameter by "failed_when" with an
additional condition. The intention was to satisfy ansible-lint. However
the task is used only to collect information and should never cause a
fatal error in Ansible Playbook. Instead, the result is properly tested
in subsequent tasks. This commit keeps the "failed_when" parameter in
order to satisfy the ansible-lint but change the condition to false.