Fix tasks in ansible_sssd_ldap_config
The 733ef1ff839a4c0a545881fb7f6b5d63439f7a9d caused an Ansible task to
fail by replacing the "ignore_errors" parameter by "failed_when" with an
additional condition. The intention was to satisfy ansible-lint. However
the task is used only to collect information and should not cause a
fatal error in Ansible Playbook. Instead, the result is properly tested
in subsequent tasks. This commit keeps the "failed_when" parameter in
order to satisfy the ansible-lint but change the condition to false.