Update CIS requirement for SSH access limit
The 5.3.4 CIS requirement for RHEL7 and 5.2.4 CIS requirement for RHEL8
and RHEL9 were updated to automated status by using the
sshd_limit_user_access rule. This rule now has OVAL and tests. There is
no automated remediation for this rule because it varies for each
system depending on site policies.