remove unnecessary --pesize from RHEL kickstarts
This is an optional optimization, setting the Physical Extent
size of the LVM Volume Group, a.k.a. the smallest atomic object.
When set properly, this can lessen the impact of LV fragmentation,
but the default of 4MB is already reasonable.

The current --pesize=4096 doesn't even change this default,
so literally nothing changes with this change.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>