remove wrong temporary password disablement in pci-dss RHEL kickstarts
This comes from

    https://github.com/ComplianceAsCode/content/pull/3660

which mentions RHBZ#1651624 (from 2018) as the reason.

That bugzilla was never valid - the original 'bootloader' line used
incorrect password syntax, likely copy/pasted from a SCAP Content
kickstart.

All that was needed was to just use the correct syntax,
so re-enable the grub2 password, using a hash identical to
other kickstarts.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>