Fix issue when adding fstab entries with iso9660
This fs type has an issue in which the entries in mtab could describe a
value for blocksize, however that's not recognized by fstab, there the
name for this same option is 'block'.
The issue can be seen from rule
mount_option_nodev_nonroot_local_partitions, if there was a cdrom
mounted manually, and then this rule was remediated. The system
wouldn't be able to reboot

Signed-off-by: Edgar Aguilar <edgar.aguilar@oracle.com>