fix: sysctl/tests: change wrong_value_d_directory.fail.sh to test all directories
And thus wrong_value_usr_local_lib.fail.sh is not needed any more.

Now wrong_value_d_directory.fail.sh tests OVAL that it finds all
relevant wrong values.

During remediation test that all relevant wrong values are fixed.

I believe testing one directory at a time is not benefical usage of
testing resources. As we want to test one item, it is to test if list of
directories implemented in OVAL or in remediation are not in sync.