OCP4 STIG: Several RBAC related controls are no longer Inherently Met, let's use a manual rule to cut duplication
This patch reflects the changes to the OCP4 STIG draft. Several rules
that used to be marked as Inherently Met with justification pointing to
RBAC are now applicable/configurable with a manual rule that asks for
the RBAC rules to be reviewed.

In order to cut down on duplication and provide the same rendered
spreadsheet output as DISA expects, I also implemented a manual rule
that covers the requirement and its text can be resued.