Include an OVAL template for ipv6 applicability
This template will be used by the ipv6 applicability to check if IPv6 is
disabled or enabled in the system. It will rely on "ipv6.disable" boot
parameter defined in /etc/default/grup as this is the most reliable
way to disable IPv6 on a system. It is not considering runtime
configuration but this could be extended in the future, if deemed
necessary.