interface change: bash_replace_or_append: key is key
'key' parameter value is treated as literal and for it there is no
transformation of any kind. Same as for 'value'.

Instead of transforming 'key' in some fashion, implement
'key_starts_with' parameter and if true, add ^ to key_regex that is used
to match the key. To keep situation as is, removed from all where ^ was
used.

Also implement 'key_regex' parameter to allow to define any regexp to
match key. '^' is not added, but 'word_boundary' is.