Update sysctl template OVAL
There are configurations set by packages in /usr/lib, so it is possible
to find there the expected configuration, but it is not recommended to
modify those files in case of a non compliant configuration. So
modified OVAL to check those files in a way that not touching them
would fix any non compliant scenario.

This means that the rule can pass if the expected conf is included in a
file in /usr/lib. But also if there is a non compliant value there, and
it is overwritten by a configuration in a different file.

Signed-off-by: Edgar Aguilar <edgar.aguilar@oracle.com>