Add a one-off tool for semi-automated component assignment
This commit adds a new script `utils/assign_components.py` which
can be used for semi-automated component assignment for rules.
The components are assigned to rules in  "obvious" cases, such as:
- the name of the component is a part of the rule ID
- the rules is in a Group that is dedicated to a specific component
- the rule is package_installed, package_removed, etc.
- the rule uses templates where we can easily infer the component based
  on the template or a template parameter
The script uses data in CSV files that have been manually improved
and they capture current state of the `linux_os` benchmark.
The script generates a YAML otuput to a given directory where each
YAML file will represent 1 component. We expect an immediate removal
of this script and CSV data.