Update CIS requirement about nftables table
RHEL systems use firewalld for firewall management. Although nftables
is the default back-end for firewalld, it is not recommended to use
nftables directly when firewalld is in use. firewalld uses the inet
firewalld that is created when firewalld is installed. The OVAL check
cannot be automated but an SCE is availble.