Add utility to render components in HTML
This commit adds a script `render_components.py` that renders
mappings from components to rules in HTML format. It's also
added to the pages that we publish online.

The goal is to enable easy sharing of component mappings
outside of the team and simplify the review of component rules
for content developers.