Add modified audit suid privilege function rule for CIS
Add a new rule to accommodate the differences between STIG and CIS
regarding execve audit in RHEL8 and RHEL9.

When analyzing a RHEL8 or RHEL9 server with the CIS Workbench after
remidiation it fails the settings of `audit execve` because
of a difference between STIG and CIS in recommended settings.