OCP4: Remove obsolete rule scheduler_no_bind_address
This rule was problematic because it applied at the same time to both
versioned platforms but also not hypershift. This resulted in having two
`<platforms>` elements in the resulting DS which means that once either
of them passed (and the not hypershift would pass on a regular install)
the rule would be evaluated even though it shouldn't be.

Because this rule only ever applies to obsolete releases (4.8 and 4.9),
let's just remove it.