Implement rules for CIS OCP Section 1.1
Now that we have a profile and control files for CIS 1.4.0, we can start
wiring up the existing rules.

This commit reuses a bunch of the existing rules we have from older
versions of the OpenShift CIS benchmark. The latest version of the
benchmark (v1.4.0) includes hardening permissions from 644 to 600 and
this change takes that into account, too.