Add audit_rules_login_events_faillock to RHEL 8 STIG
The selection of the rule audit_rules_login_events_faillock
has been commented out with a comment that the RHEL-08-020017
needs to be implemented first. But, the RHEL-08-020017 is now
implemented by rule accounts_passwords_pam_faillock_dir, which
didn't exist before. Therefore, we can uncomment the rule
selection.

The rule "accounts_passwords_pam_faillock_dir" sets the value
of the `dir` configuration option to `/var/log/faillock`
which is the value of the `faillock_path` product property.
So, we only need to use the product property in the rule.yml.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2167999