Check only local users home directories
The rule `accounts_user_interactive_home_directory_exists` states
in rationale that we need to ensure that home directories of
interactive users exist. The text indicates that this requirement
is relevant to local interactive users.

However, the current implementation of the OVAL check uses
the `unix:password_object` element, which also returns non-local
(LDAP) users, because the implementation of OpenSCAP makes use of
the `getpwent()` system call, which browses all users provides by
the NSS.

In this commit, we will change the implementation so that only
local interactive users will be considered. We will achieve this
by parsing the data directly from `/etc/passwd` using the OVAL
`ind:textfilecontent54_object` instead of using the `unix:password_object`.

Also, the rule description is clarified.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=2203791