Allow 0700 permission for UEFI files
This rule was asking 0600 permission for user.cfg file even in a UEFI
boot partition. However, UEFI usually uses a VFAT file system, which
makes the chmod command ineffective if the filesystem "umask" mount
option is set to "0077", as it is by default. If the permissions of
files in /boot/efi using VFAT file system need to be changed, the umask
value should be updated in fstab. In addition, CIS allows 0700
permission for /boot/efi and currently it is the only profile using
this rule. This PR makes 0700 permission accepted.