Refactor Ansible remediation
The used approach was not efficient on detecting relevant local
partitions and the performance of the find module was not optimized
since it would search some directories multiple times. The Playbook was
refactored to efficiently detect local partitions and locate
shosts.equiv files on them. Note that this Playbook is sharing many
code with the dir_perms_world_writable_root_owned rule which was also
refactored by me some days ago. In a separate PR I indend to move this
logic to a macro and avoid code duplication.