Skip OVAL schematron validation in CI
Stop performing OVAL schematron validation in our GitHub Actions CI.
The additional validation takes a lot of time and slows down tests
significantly. OTOH the value that it adds is low. Most of the problems
will be discovered by XML schema validation which we keep in the tests.
Advanced problems are discovered in SCAPVal runs that are also a part
of our daylife.