Update rule description of display_login_attempts
This rule demands the pam_lastlog.so module to be properly configured
in order to show the previous successfull logins. This is already
implicit in login program and configurable in ssh and other programs
which uses PAM, but it is not ensured in other programs such as gdm*
and su*. To complement program whitout a native option to show the last
login, the pam_lastlog.so module is used. This is extensively
documented in https://github.com/ComplianceAsCode/content/issues/9031