Refactor fix_rules.py
Change callback functions which name starts with `has_` so that these
functions won't load YAML files inside them but instead they will
consume a loaded YAML as a parameter.  This will separate the logic from
YAML parsing allow us to reuse these functions in a different context in
future. The functions are used as a callback therefore they all need to
have the same interface but that can lead to unused parameters in some
of the functions.