OCPBUGS-11696: Update encryption type to support 4.13 deployments
The latest CIS guidance allows OpenShift users to configure the
encryption type for the API server to `aescbc` or `aescgm`. Previously,
we only evaluated `aescbc`, but this commit updates the check to also
check that `aesgcm` is acceptable. It also updates the remediation to
use the variable for the remediation, and clarifies its usage.