Update test scenario scripts in alignment to the check
In the context of pam_lastlog and postlogin or login files, the control
is not so relevant and don't need to be tested since it can depend on
other lines, such as conditionals. In fact, it is important to check if
the "showfailed" is present without the "silent" in the same line.