New Bash macro to configure pam_lastlog.so
In systems using authselect, there is a feature to properly configure
the pam_lastlog.so module in order to be silent or to show the failed
logins. It was created a macro to use this authselect feature whenever
available. The Bash remediation was also updated to use this new macro.