Build profile bash scripts differently
This commit introduces a new script to generate profile oriented
Bash remediation scripts. It replaces the current way that uses
the `oscap` tool under the hood. The new script doesn't use `oscap`.
It processes the data stream directly and it generates the remediation
for all profiles at a single pass. The new script is faster than the
current script by 80 %.