Create functions to collect prometheus metrics
The goals is to collect prometheus metrics from all control files used
by the informed products. Therefore, the function iterates with the list
of products, locate all profiles files and search for controls used in
profile files. If a control is found, a set is incremented to further
usage. Once all relevant control files are identified, stats are
collected from each specific level in a policy. This commit cares of
collectiong the necessary information. Next commits will implement the
Prometheus format.