Remove wrong OVAL check
The OVAL check in configure_opensc_nss_db was wrong and unreliable. It
was expecting to find the opensc module information in pkcs11.txt file
but it is not expected to find this information there. This information
should be properly checked by consulting the NSS database through proper
commands. Currently, there is no file to be consulted and no OVAL check
for NSS database. Removing this OVAL is the correct approach.