Fix Ansible Tasks order
In https://github.com/ComplianceAsCode/content/pull/11033, we
have switched to a new script for generating profile oriented
Ansible Playbooks. Unfortunately, when Python 2 is used the
generated Ansible Playbooks don't preserve the order of Ansible
Tasks in the order defined in the SCAP source data stream.
The wrong order of Ansible Tasks in a Playbook might cause
an unexpected conflict between them during the run, for example
https://github.com/ComplianceAsCode/content/issues/11104.
The root cause of the problem is that dictionaries in Python 2
don't preserve order of elements but starting from Python 3.6
the dictionaries preserve order of its elements.

Fixes: #11104