Using an empty value for group has no effect
The default option in this variable used by pam_wheel.so was an empty
string. However, using an empty string in pam_wheel.so group option has
no effect. The result in this case would be an invalid configuration
to achienve the desired goal. It was included a generic group as
default value.