Rephrase the rule description and include warning
Make it clear that this rule does not configure pam_wheel.so but only
ensures the group expected to be used in pam_wheel.so exists and has no
members. The warning is referencing the use_pam_wheel_group_for_su rule
which in fact configure pam_wheel.so group option.