Review and update pcidss_4 requirement 2.2.4
PCI-DSS is not specific about which services must be removed or disabled
and it is not restricted to server systems. Therefore, in this
requirement basically only rules related to obsolete and unencrypted
services were selected to remove the respective packages while some
rules related to known non-essential services were selected to disable
the respective services.