Fix test scenarios for fapolicy_default_deny
The test scenarios were using a macro which was enough to test the OVAL
but was breaking the fapolicyd service due to syntax error. The same
macro was used twice while the first call was unnecessary. This commit
removes the unnecessary call and replace the macro by another that does
not break the fapolicyd syntax.