Adopt gpasswd command for Bash remediation
Thanks @teacup-on-rockingchair for suggesting this command.
Also used the command in group_without_members.pass.sh test scenario.
It is intentional to not update the Ansible remediation since it would
demand to use the command module. But the gpasswd command doesn't have
different return codes to know when members were removed or not from the
group. It would make the Ansible task to be reported as changed always
or never, which may cause confusion.