Include test scenario for extra rules
The logic implemented in OVAL already prevent failures when there are
more audit rules than privileged commands in the system. One valid case
is when a package including privileged commands is removed from the
system. The audit rule will remain there, but the commands are no longer
present in the system. This is a valid case and the check should not
fail. It was included a test scenario for this case.