Review and update pcidss_4 requirement 8.3.6
The requirement only requires a minimal length of 12 characters
including both numeric and alphabetic characters. It is recommended as
good practice to use special characters and a combination of upper and
lower cases but this is not required. Therefore, only required rules
were selected while recommended rules were moved to related_rules.