Return empty list when parameter is not found
`get_parameter_from_yaml()` doesn't handle KeyErrors.
This changes the function to return an empty list when the section is
not found.

It could also return `None`, but the expected return type is list.

Profiles with no `selections` are extending another profile,
let's not count the control usage multiple times.