Adopt new macro in file_permissions_unauthorized_suid
Updated the OVAL to use the create_local_mount_points_list macro.
This commit completes the improvements in
file_permissions_unauthorized_suid rule. To have some idea of the gains,
in a testing VM was created about 500.000 files and executed the OVAL
check before and after the improvements. The assessment time was reduced
from 2:25 to 1:45.