Fix ubuntu OVAL for accounts_passwords_pam_faillock_unlock_time
The ubuntu OVAL was incorrectly checking for the condition:
- unlock_time not present in pam.d/common-auth
OR
- unlock_time present in /etc/security/faillock.conf

This means that the OVAL passed when unlock_time was not
present in either of them.

The fix is analogous to the implementation in shared.xml,
where the OVAL checks for presence of unlock_time in either
one of pam.d/common-auth or faillock.conf, but not both.

The fix results in passing test ubuntu_empty_faillock_conf.fail.sh