Update test scenario to test file hashes
The intention is to modify the hash of a file managed by rpm package.
The test was searching for specific file name in /usr/share/doc to be
modified but some systems might not have the files assumed to be
present. Therefore, the test scenario was updated to modify the ls"
command, which is present even in minimal installations and a small
modification for testing purposes is not harmfull for the system.