OCPBUGS-20015: Add remediation for RHCOS banners
Public entities typically require systems to present users with a
specific banner when they log in. We have rules that check for a banner,
and a remediation in the description, along with a remediation we use in
testing.

This commit include the same remediation as a formal remediation, which
can be applied through the operator, instead of requiring users to
copy/paste them from the check result or rule description.