remove the task which deletes artifacts from automatus GH workflows
we are experiencing problems after upgrade of delete-artifact action to version 4 and they are degrading our ability to use CI efficiently.
This will be probably only temporary fix before we figure out what is the problem with the GH action and its usage in our repository.
At the same time removing this task should not pose any serious problems, old artifacts are deleted after 90 days.