sysctl template: allow skipping of runtime checks
When `check_runtime` is false the template will not generate checks
for the runtime configuration. By default the runtime checks are
always generated.

The goal of these rules is to check for the sysctl both disk and
runtime configurations. But the way CO checks for runtime sysctls may
not be the most reliable.

On OCP some sysctls are not represented with fidelity on the openscap
pod, despite being correctly configured and in effect in the
underlying node OS.