Add stigrefs after references from controls
Move the step of adding a stigref to a later stage in the process,
specifically from the moment rule.yml files are loaded to the time when
references from controls are already added to rules.

This change will allow us in future to use control file as the source of
stigids. Up until now we can't do it because adding stigrefs depend on
existence of stigid key in references in the rule object. If we want to
add stigids from control files we need to add stigrefs after stigids are
added, which means after controls are processed instead of during
original rule.yml files.