Flush automatus test logs before outputting results
This ensures that logs contain all the test stdout/stderr outputs
at the moment of automatus.py announcing their results on its output.

This is useful for an external automatus.py wrapper, which parses
its output in real-time, while also processing the logs in real-time.

The flushing is done here and not deeper, because it seems unnecessary
to flush after every 'retry' attempt, only when the test is done
running and its result is about to be announced by automatus.py.

Signed-off-by: Jiri Jaburek <comps@nomail.dom>