Unify options naming
The option `--missing-stig-ids` is actually checking the references
that are represented using `stigref` key in references in resolved
rule YAML files. The naming of the option makes a confusion with
`stigid` references. Therefore, we will rename the option to
`--missing-stigref-refs` which will align this option name with
the other options of this script, eg.` --missing-anssi-refs` is
checking for missing `anssi` references in rules so
`--missing-stigref-refs` will check for missing `stigref` references
in rules.