Extend the stable-profiles test
We will add the following profiles to the stable-profiles test:
- RHEL 7: CIS, PCI-DSS, STIG
- RHEL 8: CIS
- RHEL 9: CIS, PCI-DSS, STIG

The test will fail the CI if a rule is added or removed from a profile.
That is especially useful for the profiles that are generated from
control files that are shared among multiple profiles or products. The
test become more relevant after the prodtype removal, because now the
inclusion of rules to a data stream is driven only by profiles.