Remove unecessary sysctl-file-exists rule
This rule ensure that a sysctl file with a specific name exists.
The rules that create this file pass by default.
Furthermore, not sure why this rule is needed, since what should trigger
the creation of the file is a sysctl configuration rule failing, not the
absence of a file.