Stabilize resolved profiles
Currently, the resolved profiles have random order of items in the
selections list. This makes it less convenient to update the profile
stability test data, because copying a resolved profile causes shuffling
the git diff. This problem is a recent regression. Git bisect shows
that it has been broken by 890ce49010529494bea0dabeedec0326fa885795.
The proposed fix is to use a list instead of set so that the original
order of elements is preserved.

Related to conversation in #11719.