CMP-2375: Implement a new rule for checking audit logging is enabled
We already have a rule that checks that audit logging is configured to a
specific level called audit_profile_set. While this works, it was being
used for CIS OpenShift controls 3.2.1 and 3.2.2.

CIS 3.2.1 is really just checking if audit logging is enabled, or that
the profile isn't set to `None`.

CIS 3.2.2 is checking that the audit profile is set to
`WriteRequestBodies`.

Using the same rule for both controls was causing confusion because
control 3.2.1 should be satisfied by default with OpenShift using the
`Default` audit profile. Control 3.2.2 is a level 2 control that
requires users to make a change to the api server configuration to
remediate the finding.

This commit breaks the rule into two separate rules so that its easier
to see that one rule is clearly for CIS 3.2.1 and the other is for CIS
3.2.2.